+++ src/Backends/SDL/GuiSDL.cpp
@@ -480,7 +480,8 @@ namespace tgui
 
         m_viewport.updateParentSize({static_cast<float>(m_windowSize.x), static_cast<float>(m_windowSize.y)});
         m_view.updateParentSize({m_viewport.getWidth(), m_viewport.getHeight()});
-        m_renderTarget->setView(m_view.getRect(), m_viewport.getRect());
+        m_renderTarget->setView(m_view.getRect(), m_viewport.getRect(),
+                                {static_cast<float>(m_windowSize.x), static_cast<float>(m_windowSize.y)});
 
         GuiBase::updateContainerSize();
     }